HUGE UPDATE
This is a completely huge update on the project. Queries 1,3,4 were removed and Q2 was slighly modified (Q5 remains the same). The removed one were replaced by new ones. CountMinSketch was implemented and this is why these queries were changed. We want in the current scope of this project, both reservoir sampling and countmin sketch to be applicable to our queries so we can compare them better. 

In short
- Q1,Q3,Q4 updated.
- Q5 stays the same, Q2 slightly modified.
- Added CountMin Sketch from spark's lib
- Replaced custom made sampling with Spark's one (this has to be reconsidered)
- Added a printmethod to show the results for all queries analytically